[Feat] #43 로그인 고도화, 회원가입, 탈퇴, 조회 제작

+++ LoveTriangle-LMS/src/main/java/com/module1/crud/users/model/service/UsersService.java
@@ -38,7 +38,7 @@ public class UsersService {
         }
     }
 
-    public boolean deleteUser(Long userId) {
+    public boolean deleteUser(int userId) {
         try {
             int affectedRows = usersDAO.deleteById(userId);
